De couple crawler from source
Now we can have as many sources as we want. Crawler is able to parse
JSON and HTML(5). The specifics of each source, stays inside the
source... source code.
