+++ lyricsto
@@ -2,7 +2,7 @@
 # frozen_string_literal: true
 
 require_relative 'menu'
-require_relative 'urlbuilder'
+require_relative './sources/azlyrics'
 
 class LyricsTo
   attr_reader :artist, :song
@@ -12,10 +12,10 @@ class LyricsTo
     @song = song
   end
 
-  def build_url
-    URLBuilder.new(artist: artist, song: song).broadcast
+  def azlyrics
+    Sources::AZLyrics.new(artist: artist, song: song).fetch_lyrics
   end
 end
 
 options = Menu.new(artist: ARGV[0], song: ARGV[1]).create
-LyricsTo.new(artist: options['a'], song: options['s']).build_url
+LyricsTo.new(artist: options['a'], song: options['s']).azlyrics
